move uploader under ingestion engine
